Support Variation

+++ script/newcow.sh
@@ -4,7 +4,7 @@
 #   Clear specified cow while the system is online. Only specified computer must turn off.
 #   This is mostly used on booting computer.
 # Usage:
-#   newcow.sh <computer_name> <tid> <image_path> <image_loop_name> <cow_path> <cow_loop_name> <cow_size>
+#   newcow.sh <computer_name> <tid> <image_path> <image_loop_name> <cow_path> <cow_loop_name> <cow_size> <cow_source>
 
 COMPUTER_NAME=$1
 TID=$2
@@ -13,6 +13,7 @@ IMAGE_LOOP_NAME=$4
 COW_PATH=$5
 COW_LOOP_NAME=$6
 COW_SIZE=$7
+COW_SOURCE=$8
 
 date >> /tmp/a
 echo $COMPUTER_NAME >> /tmp/a
@@ -22,6 +23,7 @@ echo $IMAGE_LOOP_NAME >> /tmp/a
 echo $COW_PATH >> /tmp/a
 echo $COW_LOOP_NAME >> /tmp/a
 echo $COW_SIZE >> /tmp/a
+echo $COW_SOURCE >> /tmp/a
 
 image_size=`blockdev --getsize $IMAGE_LOOP_NAME`
 
@@ -30,7 +32,11 @@ sleep 1.5
 dmsetup remove $COMPUTER_NAME
 losetup -d $COW_LOOP_NAME
 rm ${COW_PATH}/${COMPUTER_NAME}.cow
-dd if=/dev/zero of=${COW_PATH}/${COMPUTER_NAME}.cow bs=1M count=0 seek=$COW_SIZE
+if [ "$COW_SOURCE" = "NULL" ]; then
+  dd if=/dev/zero of=${COW_PATH}/${COMPUTER_NAME}.cow bs=1M count=0 seek=$COW_SIZE
+else
+  cp -a ${COW_PATH}/${COW_SOURCE}.cow ${COW_PATH}/${COMPUTER_NAME}.cow
+fi
 losetup $COW_LOOP_NAME ${COW_PATH}/${COMPUTER_NAME}.cow
 dmsetup create ${COMPUTER_NAME} --table "0 ${image_size} snapshot $IMAGE_LOOP_NAME $COW_LOOP_NAME p 64"
 ietadm --op new --tid $TID --lun 0 --params Path=/dev/mapper/${COMPUTER_NAME}
